+++ changelog.md
@@ -48,6 +48,8 @@ Notice: All 1.7.x changes are present in 2.0.x aswell
 - Join tables will be no longer be paranoid (have a deletedAt timestamp added), even though other models are.
 - All tables in select queries will now be aliased with the model names to be support schemas. This will affect people stuff like `where: {'table.attribute': value}
 
+# v1.7.6
+- [BUG] fixes issue where primary key is also foreign key [#1818](https://github.com/sequelize/sequelize/pull/1818)
 
 # v1.7.5
 - [BUG] fixes bug with some methods relying on table information throwing strange errors [#1686](https://github.com/sequelize/sequelize/pull/1686)